backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.0
SQLAlchemy==2.0.23
pydantic==2.4.2
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
aiofiles==23.2.1
jinja2==3.1.2
pytest==7.4.3
celery==5.3.4
pymongo==4.5.0
requests==2.31.0
python-multipart==0.0.6
SQLAlchemy-Utils==0.41.1